Added rating count custom field and command
It is now possible to count the amount of ratings a product has and save the value in a custom field that is synced to BI. The field is updated when a review is created, edited or deleted. There is also a command that updates all products with reviews at once.